extract StorageModule
